feat: add SystemWebShim to enable some binary compat (#14)
This will allow libraries built with System.Web for framework to have a chance at working on core. This will purely type forward, but these assemblies would be built assuming everything to be in System.Web.dll, so this will allow them to at least be able to find the types that are supported.