Fixing Argument out of range exception for HtmlTextWriter Style and pushing HtmlTextWriter webformsfeed (#50)
While writing attributes a check is done per index as displayed
https://github.com/CoreWebForms/CoreWebForms/blob/main/src/HtmlTextWriter/CssTextWriter.cs#L282
which was causing the Out of range exception. The reason for out of
range exception while porting we change the data structure to List
(which size is always 0) instead of array as here
https://referencesource.microsoft.com/#System.Web/UI/CssTextWriter.cs,28
. I have changed to keep as array as there is not much performance or
memory gain.

Also made the HtmlTextWriter.csproj as shipped to be available in
webformsfeed.