Enable compiler to reuse dependencies (#94)
This changes the SystemWebCompilation to track compiled pages for its
lifetime so that items can be reused.

As part of this, there is now the ability to queue a page for
compilation if it is a dependency of another (i.e. a user control may be
required for a page, or a user control in a user control). In these
cases, each dependency is compiled to its own dependent assembly. In the
case of dynamic loading, each of these will be loaded into its own
AssemblyLoadContext so they can be unloaded.