Add a source generator to use C# 12 interceptors (#102)
This will create an auto-generated, strongly-typed, method to register
all the routes. Although we don't currently support AOT, this at least
makes it so we don't need to use reflection to load up the pages if
opted into.

This change will also make it easier to set up some automated testing of
the samples as they will be published and can be run independently of
the repo.