Ensure Respone.End throws during page execution
By design, System.Web adapters do not throw when Response.End occurs, but Pages expect that to occur to stop interaction. An example where this expectation fails is during Response.Redirect calls and the error is difficult to understand. This overrides the End() behavior for the duration of the page execution to ensure the behavior is closer to framework.