Use a preprocessor for ajax files
On the original WebForms builds, there was a preprocessor that handled
rewriting c-style preprocessor directives. We don't have access to this,
and I had attempted to just comment them out. However, that didn't work
for multi-line #if DEBUG statements.

This change introduces a custom rollup plugin that will take the
original inputs and process them so we can comment out (or if we want
remove) the debug directives, including multi-line ones.