Enable assemblies to declare controls with TagPrefixAttribute (#141)
This is an attribute that was used ASP.NET to mark assemblies as
controls. We can use this to enable more seamless control compilation
both dynamically and statically.

As part of this, a new interface ITypeResolutionService is implemented
for both compilers to retrieve controls at runtime. On ASP.NET, this was
used for just in the designer, but it makes sense to use this (at least
for now) to surface the information.

Fixes #116