Use a preprocessor for ajax files (#136)
On the original WebForms builds, there was a preprocessor that handled
rewriting c-style preprocessor directives. We don't have access to this,
and I had attempted to just comment them out. However, that didn't work
for multi-line #if DEBUG statements.

This change introduces a custom rollup plugin that will take the
original inputs and process them so we can comment out (or if we want
remove) the debug directives, including multi-line ones.

The js/jsa files that are touched (besides the preprocessor.js plugin)
are all a recopy
of the original ajax files without the commented out `#if DEBUG` kind of
sections

Part of #9