Add a banned API for drawing types (#150)
We're ok using System.Drawing.Primitives (where color is), but
System.Drawing.Common is not available on all platforms. However, the
sdk seems to allow us to reference it and then it breaks at runtime.
This adds a BannedSymbols.txt so that we can add to it for the APIs we
find that are using that assembly; unfortunately, we can't just disable
all System.Drawing, as then color is not allowed.