Consolidate GetReferencedType method (#235)
We introduced ICompiledTypeAccessor as a way to get at a compiled type
given a virtual path earlier this week. In order to consolidate the
GetReferencedType code to utilize this fully, some changes to MasterPage
were required. All of these revert the code back to what the original
system did; that is, these changes had been put in as we bootstrapped
the system and now we can use the original patterns.

Fixes #232