Use an abstraction to write compilation results
This change introduces a ICompilationStrategy that produces the streams that are used for compilation as well as handles error handling. This allows the compilation system to be decoupled a bit so that we don't need to subclass SystemWebCompilation but can just pass into it an implementation of ICompilationStrategy that makes sense for the use case (i.e. dynamic or static).