Use same assembly name for both compilation and file name (#306)
Fixes #305

Currently, the compilation uses an assembly name derived from the name
of the type for the page, while the assembly name for the file name is
derived from the page's default route. Because these names are
different, an error occurs attempting to load the assembly, which
expects the compiled assembly name to match the file name.

This pull request attempts to remedy this by using the same assembly
name for the file name that is currently being used for the compilation.