Update target framework and extend type forwarding
This commit updates the project to target .NET 8.0, which involves a transition from `netstandard2.0` to support modern runtime features and enhancements in performance. In addition, more `TypeForwardedTo` attributes are added across multiple classes in the `Shim.Extensions` and `Shim` projects, broadening the range of types that are forwarded, thereby improving compatibility with ASP.NET Web Forms components and related configuration sections.

1. Updated `src/Compiler/Compiler.csproj` to use `<TargetFramework>net8.0</TargetFramework>`.
2. Added several new `TypeForwardedTo` attributes in `Forwards.cs` files, enhancing type forwarding for web control structures and profile management within the ASP.NET ecosystem.