Loading referenced assemblies to current app domain (#389)
When EntityFramework being added to the sample project, the compilation
for dynamic_page failed as it couldn't get EntityFramework related
assemblies, although those are added as direct dependency. During
Assembly load event, the EF dll is not getting loaded to the current
appdomain, so forcing referenced assemblies load which are part of
current application domain.

Co-authored-by: birojkumarnayak <birojkumarnayak@gmail.com>